[BUGFIX] Fix preview with extendToSubpages
A problem in combination with extendToSubpages still existed and
resulted in page displayed with a 403 error, if logged in the
backend, if

- current page is hidden=0
- current page has extendToSubpages=1
- parent page has hidden=1
- parent page has extendToSubpages=1
- user is logged in Backend and opens pages via preview
- admin panel is not active

The problem did not exist if the current page had extendToSubpages=0,
which was introduced while fixing issue #17599.

This is now fixed: In PreviewSimulator::checkIfRootlineRequiresPreview
it was checked if the current page has extendToSubpages=1, but this
is completely irrelevant to the visibility of the current page.
So the check is removed.

Resolves: #106336
Related: #17599
Related: #92675
Related: #92286
Releases: main, 13.4, 12.4
Change-Id: I720edc81ff29e06376252320a695f8d22def3f62
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88537
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>