[BUGFIX] Workspaces: calculate paging on correct data structure
The method `GridService->calculateStartWithCollections()` relies on
`$this->dataArray` having numeric indices, which is not the case at this
point. The method is called by `GridService->getDataArray()`, which takes
care of this, unfortunately too late.

The key conversion is now done as the beginning of the method, fixing
the pagination.

Resolves: #106308
Releases: main, 13.4, 12.4
Change-Id: Ic9b90cdeb984d01d30fbb46c27f8a4c2394dc39f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88547
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>