[BUGFIX] Fix race conditions in childnode querying custom elements
Children (or even succeeding siblings) of custom elements are not
guaranteed to be available by the time the element is added to the
DOM — i.e. when the connectedCallback() is triggered by the
browser, even when the components and their children are server
rendered, slow connections and streaming can cause initialization
race conditions. That means we must never assume that DOM
is fully loaded in connectedCallback. See #102550 for details.

In previous attempts (#102550, #102603) we made use of
HTML event bubbling and slotchange events to detect children, but these
solutions are profound and in parts fragile to master.
Therefore we use the DocumentService sledgehammer to make sure
we do not race during element initialization.

Releases: main, 13.4, 12.4
Resolves: #106371
Related: #102550
Related: #102603
Change-Id: I0f191c31d6ca029f9f2d23adccdaf0dddb568c35
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88637
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>