[BUGFIX] Allow empty config.doctype as HTML5
A specific change for HTML5 conformity was
added to the main branch but was leftover in
v12 branch as PHP 8 conformity in #102297.

The main change is that TYPO3 needs to understand
that "config.doctype = " and config.doctype as
not set is added as the same.

This is a low-level drive-by fix in v12,
and already adressed via the Enum in v13+

Resolves: #106449
Releases: 12.4
Change-Id: I631d69cb3272610552618e958ccadf35bc38c298
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88813
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>