[TASK] Make container hosts IPv4 addresses manageable with GitLab
This change replace hardcoded hosts ipv4 override in
`Build/Scripts/runTests.sh` with a variable to allow
managing it as GitLab variable in the repository for
test execution.

Upgrading runner to podman 5 is planned and hopefully
makes this workaround obsolete, but having it in place
for quicker adjustment does not hurt as intermediate
solution.

Resolves: #106463
Releases: main, 13.4, 12.4
Change-Id: Ibefe17b46cb4fe09ba4b41db6ddc765b019507bb
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88846
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>