[BUGFIX] Count elements in clipboard correctly when pasting all
In the record list, the method generating the clipboard paste
confirmation along with the "Paste in clipboard content" counted in a
wrong fashion since #106032, as the amount of records is not handed in
anymore.

Since that single method cannot determine what to count without context,
a new internal enum `CountMode` is introduced that instructs how
elements in the clipboard should get counted.

Resolves: #106474
Related: #106032
Releases: main
Change-Id: I5538d477884ec71db2fae5971d39ef64cad475bf
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88881
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>