[TASK] Adjust `EnumType` and `SetType` type schema handling
With `doctrine/dbal 4.2.0` the `enum` has been added directly
in DBAL, implementing it slightly different than the original
TYPO3 implementation, which has been mitigated with #105279
and adopted for the `set` type in advance with #105294.

TYPO3 still needs to adopt the options reading existing database
schema provided by a trait to the extended platform classes and
missed to pass the `enum` and `set` option values as `values` to
the `Column` objects instead as `platformOption` key, hidden by
the fact that getting the array representation of a column merges
the platformOptions on the same level with the direct value options.

Upcoming `doctrine/dbal` introduces more phpdoc-block array shapes,
narrowing down possible (allowed) array keys for `platformOptions`
and failing with phpstan in core tests while not beeing a technical
issue.

This change adjusts related code within the `ext_tables.sql` related
`TableBuilder` and the platform trait reading the values and setting
it to `Column` objects as value property instead of `platformOption`.

Considered as task and not as bugfix due to the fact that this is
only static code analysis related and not technically breaking.

Minor left-over in the `SetType` implementation for `doctrine/dbal 4.0`
is removed along the way.

Tested and extracted from doctrine/dbal 4.3.x test,
see: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88596

[1] https://github.com/doctrine/dbal/pull/6536
[2] https://github.com/doctrine/dbal/releases/tag/4.2.0

Resolves: #106486
Related: #105294
Related: #105279
Releases: main, 13.4
Change-Id: Ib40d8d684fdce840e4050387416227d56e39698b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88905
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>