[BUGFIX] Determine correct record uid from single record clipboard
TCA group fields offer to paste elements from the currently active
clipboard, which may come in handy. However, due to a misunderstanding
how the clipboard stores its data, this feature is broken for clipboards
that are in single record mode (aka "normal" clipboard), as the "record
uid" is a simple flag in this case.

The code is adjusted to take the array keys of the clipboard items into
account only to fetch the uid from the keys. Also, the code is cleaned
up a bit to avoid duplication.

Resolves: #92292
Releases: main, 13.4, 12.4
Change-Id: I0b482055ca0d2dee4d2278242be5717a793efb31
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88960
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>