[TASK] Prepare for a breaking `doctrine/dbal` deprecation
Doctrine DBAL 4.3.x development branch contains a deprecation
breaking TYPO3 Database Analyzer handling [1], which has been
tested and reported as breaking [2].

While the discussion has been opened and the breaking behavior
may be mitigated, this change aims to prepare for the possible
case that this may not happen and having a solution in the next
TYPO3 release in place + the ability to use it as composer patch
for existing instances not being able to update to the version
including this mitigation code.

Note that the deprecation needs to be addressed in a bigger scope
to prepare for Doctrine DBAL 5.x upgrade which includes a rework
and reconception of the full `Database Analyzer` stack handling
and out of scope for now.

Tested and extracted from `doctrine/dbal 4.3.x` test change,
see: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88596

[1] https://github.com/doctrine/dbal/pull/6831
[2] https://github.com/doctrine/dbal/issues/6880

Resolves: #106495
Releases: main, 13.4
Change-Id: Id0e3569fead208d0e2885d40316f2c358c6ddc1a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88940
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>