[BUGFIX] Fix end of tag names in parseFunc
The HTML specification allows tag names to end in spaces, line feeds,
character tabulation and form feeds. The current parseFunc only splits
at spaces. If an allowlist of tags is used, this can trigger an allowed
tag to be denied (or a denied tag to be allowed).

This commit implements all the characters that can split attributes from
tag names.

Resolves: #106548
Releases: main, 13.4
Change-Id: Icba0c262f59c7e356f22ca76b6e172168fede6c2
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89088
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>