[TASK] Improve view of redirects to disabled pages
If a redirect redirects to a page, the icon of the page is now
shown before the page id. This helps to identify problems which
occur if the page is disabled and the redirect would not work.

The internal `EXT:redirects` ViewHelper is modified to return
the page record instead of the pageId and renamed to reflect
the changed return value. Template is adopted to the new name.

Resolves: #105412
Releases: main, 13.4
Change-Id: Ia512dcbae8c2573dbde701e5b43a6417c7f8a921
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89059
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>