[BUGFIX] Properly calculate the effective PID for page translations or versioned pages
When using the marker ##CURRENT_PID### inside FormEngine, TSconfig
etc. FormEngine calculates the so-called "effectivePid"
which points to the parents' page UID, the PID. When
editing the actual page on the root-level, it's the same.

This is relevant when editing page translations, or when
editing a page in a workspace, because there was a bug:

While migrating from pages_language_overlay we forgot to
actually adapt the effectivePid to the proper record.

- For pages in default language (live), it's the "uid"
- For page translations (live), it's the "l10n_parent"
- For pages in default language (workspace), it's the "t3ver_oid"
- For page translations (workspace), it's the "l10n_parent"
  which also points to the LIVE record of the default language.

Workspace hint: This was not easy to deal with until v10, while we
still had pid=-1 records for versioned records.

Language hint (although not directly related here): the PID
and the sorting parameter of a page translation is also the same as
the PID of the page in the default language.

Resolves: #102069
Releases: main, 13.4, 12.4
Change-Id: I868e1f6c65fbbc9331c8e9117797f32cda23ac7d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89075
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>