[BUGFIX] Prevent file selection via name link when denied by IsFileSelectableEvent
When a file is marked as non-selectable using IsFileSelectableEvent, the
file browser correctly removes the '+' icon and checkboxes. However, the
file could still be selected by clicking on the file name.

This patch ensures that when a file is not selectable, the attribute
`data-filelist-action` is not set, and the button is instead disabled.

Resolves: #106560
Releases: main, 13.4
Change-Id: Ia0fcf01977ed372d64e1a996bc36ab97f27f6964
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89064
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>