[TASK] Always use the current workspaceId for WSOL in DataHandler
There might be scenarios where DataHandler gets initialized
with a different BE_USER, that's why it can be handed in.

However, when calling BackendUtility::workspaceOL()
it should always determine the workspaceId of THAT user.

This is not possible via BackendUtility::getRecordWSOL(),
so it is really bad to use this in DataHandler context.

For this reason, we use

  $row = BackendUtility::getRecord($table, $uid);
  BackendUtility::workspaceOL(
    $table, $row, $this->BE_USER->workspace,
  );

In the long-term we should either add a new parameter
to the method getRecordWSOL() or deprecate it.

Resolves: #106620
Releases: main, 13.4
Change-Id: Id3278dcfa072c0ea7777f607f5b315e366149edd
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89077
Reviewed-by: Oli Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Oli Bartsch <bo@cedev.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Stefan Bürk <stefan@buerk.tech>