[TASK] Add tests for localization l10n_mode=exclude
This is a leftover from adding tasks 5 years ago,
where this use-case was missing and not added yet.

There is an interesting side-effect in workspaces:
Because the localization happens in the WS, the default
language needs to be created as well. l10n_mode exclude
means that it modifies all languages - thus creating
a regular "t3ver_state=0" record as well (= 2 versioned records).

When "Publish Workspace" (publishAll) is used, the
modified default language version is overwriting the live
record. This is a nasty issue, because if the live record in
default language might be modified during that time, it
might be overwriting live changes by accident.

Resolves: #79244
Releases: main, 13.4
Change-Id: Iba265775fbbb5147c5663fc8767c01d6fb4ef19f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89093
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>