[BUGFIX] Load `draggable-resizable-element` in top-level scope
The image cropping functionality allows to define so-named "focus areas"
to hint editors to respect a certain area while cropping pictures.

This functionality broke with #100948 in Firefox, as the web component
`@typo3/backend/element/draggable-resizable-element` is loaded within
the content frame context, but used in a modal window, which resides
in the outer backend scaffold context.

Explicitly loading the web component in top level is not enough, as
exports of said component are already loaded and used in
`image-manipulation`.

To solve this, several changes are done:

* the class `Offset` has been extracted into a new file and is now
  independently loaded in both affected modules, which avoids premature
  loading of `draggable-resizable-element` and therefore fixes the
  scoping in Firefox
* the consumers of `initFocusArea()` have been made async as well

Resolves: #103972
Related: #100948
Releases: main, 13.4, 12.4
Change-Id: Ib0757651de478102f5067331ca4251b41380b9bb
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89140
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>