[BUGFIX] Execute `bulkInsert()` in chunks for larger datasets
The `Connection::bulkInsert()` method invites to pass bulk of
data to insert into the database, using database placeholders
to provide proper value handling. However, handling for more
then one specific database system is not easy and each single
system defines a different maximum placeholder limit and lead
to exceeding that limit with a lot of data rows passed in.

To mitigate this issue a maximum data rows limit is calculated
based on the columns count and the maximal placeholder limit,
with a safety margin. The result is used to build chunks out
of the passed insert data rows and splitted into multiple calls.

This mitigates "Too many parameters" database errors inserting
larger datasets to some degree, not fully solving it.

Extensions or instances using the `bulkInsert()` method still
recieving the aforementioned error should implement chunking
of the insert data and call this method mutiple times, which
should be more an edge case after this change.

Resolves: #106579
Releases: main, 13.4
Change-Id: Id1d14b17f0b9f506176e7734f582e80396ddca4c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89144
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>