[BUGFIX] Retain navigation toggle state when switching modules
The navigation component (aka "page tree") has a long standing issue
that the toggle state is not kept and would automatically expand again
if a backend module is opened.

When opening a backend module, the CSS class for expanding the
navigation is not forcefully set anymore, allowing to keep the toggle
state when switching modules. Instead, the navigation is forced to be
expanded when rendering the backend.

A related CSS bugfix binds the visibility of the navigation component to
the class `scaffold-content-navigation-available`.

Resolves: #92545
Releases: main, 13.4
Change-Id: If5ee559e334eda6d16b4dcdc5dffca65716f7dfd
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89209
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>