[BUGFIX] DataHandler hard delete may delete too much
A recycler related issue slipped through with #106382:
When a soft-deleted localized page is hard-deleted, which
is only available using ext:recycler, records on the
default language page and records of other localizations
are deleted as well.

The fix is to fetch soft-deleted records as well in the
code that tries to find out if the page is a localized
page or not.

Functional tests lack hard-delete scenarios, which is why
this issue has not been noticed. This case is now covered.

Resolves: #106658
Related: #106382
Releases: main, 13.4
Change-Id: Ieab1d3ff74b24d5a4aadc40a60b5b570f235ad04
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89316
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>