Revert "[BUGFIX] Strip only existing prefix in PathUtility::stripPathSitePrefix()"
This reverts commit e5b30f1a3136a51ded87f65dea088ecb38081b22
as the bugfix revealed that quite a number of usages relied
at least partly on the behaviour, documented with #106690 and
a try to fix some of them [1] which did not pay out.

That gives us time to replace consuming places with suitable
replacements and drop the internal method in the end at all.

[1] https://review.typo3.org/c/Packages/TYPO3.CMS/+/89376

Resolves: #106690
Reverts: #106605
Releases: main, 13.4
Change-Id: If7d845d8be45e8fcc13e275fe0ebcaf457077d00
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89381
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Kevin Appelt <kevin.appelt@icloud.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Kevin Appelt <kevin.appelt@icloud.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>