[BUGFIX] Fix local configuration GUI with numerical map keys
The GUI-managed map/list local configuration handling introduced
with #106553 throws an exception when entering a numerical map
array key.

The "map" type is not yet utilized in the current TYPO3 core
DefaultConfiguration, thus this bug has no impact yet.

This patch adjusts the htmlentities() call to cast
numbers to a string to make numerical keys possible.

Note that later on, numerical array keys are actually
persisted with their proper numerical array key, not a string.

Also note that the GUI submits map keys sorted with
integer numbers first, and later handling also sorts
the whole array by their key.

To prevent the GUI from reporting multiple "Write configuration"
submits as having changes, the comparison of the old and
new values no longer report changes in sorting as a difference.
Only when values or keys change, the GUI reports it as a change.

Resolves: #106704
Related: #106553
Releases: main, 13.4
Change-Id: Id5626101f0d41a15880b9ba8c808cdcdd9bc39c4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89400
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Thomas Hohn <tho@gyldendal.dk>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>