[TASK] Provide default icon for plugins
Since the synchronization of CType item groups and the wizard groups
all plugins appear in the "plugins" group of the
NewContentElementWizard. Since then, a lot of plugins appear without an
icon, because it wasn't really necessary to provide one for old
list_type plugins. This change adds a default generic plugin icon for
plugins registered in ExtensionUtility::registerPlugin.

The doc comment is adjusted to signalize that null is still a valid
value.

Resolves: #106717
Related: #102834
Releases: main, 13.4
Change-Id: I08822db08f46c0851d1e9d8b7361c97de81596e4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89421
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>