[BUGFIX] Improve visibility of hidden inline/file references
Previously, the entire row—including controls was rendered with
opacity 0.3 when a reference was hidden. This caused confusion for
editors, as it appeared that controls were disabled even when they were
not.

This change limits the opacity effect on the panel header content only,
ensuring that controls remain fully visible and clearly usable unless
explicitly disabled.

Resolves: #106744
Releases: main, 13.4
Change-Id: I2a1806b29079a3e2bc707585097638ae0e58a942
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89595
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oli Bartsch <bo@cedev.de>
Reviewed-by: Oli Bartsch <bo@cedev.de>