[BUGFIX] Fix "stringlist" site set settings not persisting values
Defining a settings.definitions.yaml file like this:

```
settings:
  foo.bar.baz:
    type: stringlist
    default: ['a', 'b']
    label: FooBar
```

shows a configuration value in the site settings editor with
a list of input fields (2 in this case with 'a' and 'b').

Changing this in the GUI did not persist the changes
properly, because validation expected a proper array
input, where a JSON input however is propagated.

This fix follows the recommendation in #106747 by
also performing the JSON decode in the `validate()`
method instead of just in the `transformValue()`
method. This is streamlined with its own
private method `decodeJsonAsFallback` to prevent
redundancy.

Can be reproduced by editing [1] and inserting:

```
  seo.sitemap.stringlist:
    type: stringlist
    default: ["foo", "bar"]
    label: StringList with default foo+bar
```

[1] EXT:seo/Configuration/Sets/Sitemap/settings.definitions.yaml

Resolves: #106747
Releases: main, 13.4
Change-Id: I82212d30f0900cdf4f6e7f3964ebfe6c73a654c0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89555
Reviewed-by: Marvin Buchmann <marvin.buchmann@typo3.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Marvin Buchmann <marvin.buchmann@typo3.com>