[BUGFIX] Fix indefinite sudo confirmation when copying backend users
With #106743 the DataHandler sudo mode access subjects (that were added
in #103252) have been extended with the record identifier as subject
context.
For new or copied record there is no UID yet, therefore a randomized
string is used, which causes issues with password grants which act
as one-time grants and therefore need an exact subject match, where
a randomized identifier causes the access subject to never match.

To solve this, the randomized NEW.+ identifiers are normalized to NEW,
in order to create stable sudo mode authentication claims, that do not
change between claim creation and claim consumption.

Resolves: #106788
Related: #106743
Related: #103252
Releases: main, 13.4, 12.4
Change-Id: I75b9b557e7ac93cfbf5764af29f5d18d48b5e56f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89616
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>