[BUGFIX] Correctly hard-delete inline foreign field children
This one was found by one of the flex->inline tests when
working on #79014 tests: When a parent record that has
more than one inline child is hard-deleted after it has
been soft-deleted (typically using the recycler), attached
children have to be hard-deleted as well, but are not.
It works when only child is attached, though.

The reason is a detail in PlainDataResolver that takes care
of resorting given child ids. Beside the situation that
resorting is probably not needed in the delete case,
processSorting() also applies the deleted restriction to
the query that is used to resort and thus "looses"
soft-deleted records.

The solution is to never apply the DeletedRestriction in
order to retain the full list of incoming ids when bringing
records into the correct order, since deleted restrictions
should have been applied already if needed - typically by
RelationHandler.

The patch resolves according @todo comments from #79014
and adds a test for a "simple" parent->inline
foreign_field scenario that does not involve flex forms.

Resolves: #106815
Resolves: #76138
Related: #79014
Releases: main, 13.4
Change-Id: I6f23ab71b78f230dbb148fa5e9fe0aa374ea19af
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89617
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>