[TASK] Remove IrreForeignFieldNonWs tests
The IrreForeignFieldNonWs tests reflect the scenario
"an inline parent is workspace aware, an inline child
is not". This has been declared unsupported in main
with #106821. The scenario is so full of issues that
the tests actively prevent fixing other issues. With
the scenario being avoided in main and any issues no
longer being fixed in v13, the tests are now removed
in v13 to unblock fixing other things. Note there is
a last comment in DataHandler mentioning the tests.
This will be cleaned up with another change since the
code covers multiple slightly different scenarios
as well.

Resolves: #106824
Related: #106821
Related: #98103
Releases: 13.4
Change-Id: Ie41d13791f7405c156b14e743c08cf9e7f192031
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89629
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oli Bartsch <bo@cedev.de>
Reviewed-by: Oli Bartsch <bo@cedev.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>