[BUGFIX] Properly discard flex relation children
This resolves an issue in workspaces 'discard' that has
been resolved for 'delete' with #106431 already: When
a record having relations inside flex forms is discarded,
attached relations are now discarded as well.

The patch implements the behavior and resolves a bunch
of todo comments within scenario csv test files that
pointed out this issue already.

Resolves: #106829
Related: #106431
Releases: main, 13.4
Change-Id: I84256189a89e9dba872012204dcee123ed661b3c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89639
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>