[FEATURE] Introduce shell auto-completion for the `typo3` script
`symfony/console` supports registering autocompletion for the
symfony-cli and console applications for several shells for
some time:

* Symfony 5.4 for Bash
* Symfony 6.1 for Fish
* Symfony 6.2 for ZSH

These are some commonly used shell variants on today's systems.

This change registers the Symfony completion command to
make it available as part of the `typo3` script. It provides
help on how to register the autocompletion lists via:

> bin/typo3 completion --help

The default autocompletion behaviour is helpful and makes
using the CLI command easier, as it reduces the need
to manually complete commands, simply by using the `TAB`
key to cycle through completions/suggestions.

With this in place, it is possible to further improve TYPO3
commands, or commands provided by extension authors, by adding
a special completion handler for command options, like looking
up values or attributes/options in a database or a file
etc. (dynamic).

Resolves: #106839
Releases: main
Change-Id: I5d8e249b905100656a7d930f233656b119ac41ac
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89659
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>