[TASK] Add SchemaTypeInformation in favor of getSubSchemaDivisorField
Schema API has a drawback with the existing
getSubSchemaDivisorField() as it does not cover
"uid_local:type" field from sys_file_reference.

This is now solved by introducing the information
in a DTO "SchemaTypeInformation" which will be added
to v13 as well, and supersed getSubSchemaDivisorField()
in v14 with a followup patch.

This way, Schema API gets more powerful in a few places
in TYPO3 Core, where this special handling of a remote
field value wasn't supported yet.

Resolves: #106850
Related: #104002
Releases: main, 13.4
Change-Id: I84abbcb7f462e6c752c9c17efc0a529793f640bc
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89685
Reviewed-by: Oli Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oli Bartsch <bo@cedev.de>