[BUGFIX] Streamline type of buttons within backend button bar
The buttons added to the backend button bar can be modified through a
`ModifyButtonBarEvent`. Buttons, which were initially added to the
button bar using the `ButtonBar::addButton` method, are sorted by
position and group. This array structure was not properly reflected in
the `ModifyButtonBarEvent`, which is now fixed by introducing a
dedicated PHPStan type for the rather complex array structure.

Resolves: #106885
Releases: main, 13.4, 12.4
Change-Id: Ia361262a1836eb56039f51d4cb60a3eccbfcc478
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89748
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>