[BUGFIX] Mitigate TypeError in FormEngine DataProvider `TcaInline`
With #104584 the one or other native types has been added to
`\TYPO3\CMS\Backend\Form\FormDataProvider\TcaInline` methods,
leading to issues in calling paths for TCA type=inline fields
with `foreign_table` configuration.

Configuration like following:

    'tx_myext_entities' => [
        'exclude' => true,
        'label' => $ll('pages.tx_academiccontacts4pages_contacts'),
        'l10n_mode' => 'exclude',
        'config' => [
            'type' => 'inline',
            'appearance' => [
                'collapseAll' => true,
                'expandSingle' => false,
                'showNewRecordLink' => true,
                'newRecordLinkAddTitle' => true,
                'levelLinksPosition' => 'top',
                'useCombination' => false,
                'suppressCombinationWarning' => false,
                'useSortable' => true,
                'showPossibleLocalizationRecords' => false,
                'showAllLocalizationLink' => false,
                'showSynchronizationLink' => false,
                'enabledControls' => [
                    'info' => true,
                    'new' =>  true,
                    'dragdrop' => true,
                    'sort' => false,
                    'hide' => true,
                    'delete' => true,
                    'localize' => true,
                ],
                'showPossibleRecordsSelector' => false,
                'elementBrowserEnabled' => false,
            ],
            'enableCascadingDelete' => true,
            'foreign_field' =>  'page',
            'foreign_sortby' => 'sorting',
            'foreign_table' => 'tx_myext_domain_model_entity',
        ],
    ],

requires `nullable integer` database fields as for the column
`tx_myext_entities` in the table as it is a counter field due
to relation handling.

`l10n_mode=exclude` means that the field is not translated and
does not have a default language row during page localization,
trying to pass array value `NULL` as 4th argument to the method
`TcaInline::resolveConnectedRecordUids()` expecting `string`,
leading to the TypeError.

This change adds a string cast combined with null-coalesce fallback
to the call as a bugfix to restore non-crashing state as bugfix,
mitigating the introduced regression based on adding native types.

Long term it should be investigated if calling the method make
sense at all or should be early prevented.

Resolves: #106889
Related: #104584
Releases: main, 13.4
Change-Id: Id6e52f7403e1773809ec12733954475c04edba70
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89734
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Garvin Hicking <garvin@hick.ing>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>