[TASK] Remove import of `RecordTableElement` in workspaces backend
The import of `RecordTableElement` in the `@typo3/workspaces/backend`
module is only required as the code assumely needs to know the exact
type of an element returned by `querySelector()`.

The queried element `typo3-workspaces-record-table` already properly
declares itself in the `HTMLElementTagNameMap`, therefore
`querySelector()` is very well aware of the element being returned.

This allows us to remove the explicit type import.

Resolves: #106900
Related: #106680
Releases: main, 13.4
Change-Id: I02bd27d33220b385ae1b9848ec6ae6ca4620ff8e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89751
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>