[TASK] Rework Live Search filtering
The filtering approach in the Live Search is refined to directly react
on user interaction when toggling a checkbox. Before, the user had to
close the filter dropdown, which is not in line with the search field
behavior as it doesn't require an explicit action.

If a user changes the search demand while a request is running, the
current request gets aborted to avoid letting the server handling
obsolete requests.

Also, to make the intention of the filter dropdown more clear, the
trigger icon is changed to an actual "filter" icon.

Resolves: #106848
Releases: main, 13.4
Change-Id: Ia8277017ee0d8fafd6e3301a52fa7cad1cc36520
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89828
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>