[BUGFIX] Check correct file location in DeferredBackendImageProcessor
The condition whether a task can be processed
by this processor involves a file existence check.
It should check, if the expected target file
for the processed file already exists within
the processed-files folder.

The check was wrong, because it used the
root processing folder instead of the nested
processing folder.

Resolves: #106904
Releases: main, 13.4, 12.4
Change-Id: Ie452ecc92108c8458674661aa8a5a8576fc115a8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90164
Tested-by: Garvin Hicking <garvin@hick.ing>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Garvin Hicking <garvin@hick.ing>