[BUGFIX] Avoid invalid events on initial render in property grid editor
With #106485 the proper "dirtyness" checks of editing forms in the
backend UI was impacted so it recognized every form as "dirty" in
case a "Email to sender" or "Email to receiver" finisher was added.
The was caused by `<typo3-form-property-grid-editor>` pushing the
`PropertyGridEditorUpdateEvent` to subscribers not only on change,
but also on initialization.

Ensure that an update event is only dispatched if the value
of the entries change, not when they are initially set.

Resolves: #106907
Related: #106485
Releases: main, 13.4
Change-Id: I4a63f110c9f5bd108be0531b89df79cd708ad39e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89756
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: Garvin Hicking <garvin@hick.ing>