[BUGFIX] Ensure FlexFormTools reset state of injected TcaMigration
With #104268, `TcaMigration` is injected as a service into
`FlexFormTools` under the assumption that it is stateless.

However, the `$messages` property actually carries state and will
unnecessarily stack output messages from further runs.

This patch ensures the message stack gets reset on each run
of the `migrate()` method so the class can act as a proper
dependency-injected pseudo-stateless service.
An added test simulates this logic.

Resolves: #106912
Related: #104268
Releases: main, 13.4
Change-Id: Ifc3425d27f491ba2d961676ef5f9c290482ea401
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89774
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Oli Bartsch <bo@cedev.de>
Tested-by: Oli Bartsch <bo@cedev.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>